Extend maximum size for durations. As long as there are errors occuring due to measurement calculating extremly big numbers the type BigInteger is sometimes not big enough

+++ benchbuild/utils/schema.py
@@ -332,7 +332,7 @@ class Regions(BASE):
                     index=True,
                     primary_key=True)
 
-    duration = Column(BigInteger)
+    duration = Column(Numeric)
     id  = Column(Integer, primary_key=True)
     name = Column(String)
     events = Column(BigInteger)
